requirements.txt
Django==1.11.4
django-annoying==0.10.3
django-extensions==1.8.1
django-mptt==0.8.7
django-mptt-urls==2.0.3
ipython==6.1.0
psycopg2==2.7.3
sorl-thumbnail==12.3
django-celery==3.1.17
celery==3.1.23